# docs/swagger.yaml
definitions:
...
  httputil.Error:
    properties:
      code:
        type: string
      details:
        additionalProperties: {}
        type: object
      message:
        type: string
    type: object
...
  tenant.Algorithm:
    enum:
    - RS256
    - ES256
    type: string
    x-enum-varnames:
    - RS256
    - ES256
  tenant.CreateTenantRequest:
    properties:
      algorithm:
        $ref: '#/definitions/tenant.Algorithm'
      domain:
        type: string
      id:
        type: string
      issuer:
        type: string
    type: object
  tenant.MFAPolicy:
    properties:
      methods:
        description: e.g., ["totp"]
        items:
          type: string
        type: array
      required:
        description: '"none", "optional", "required"'
        type: string
    type: object
  tenant.SigningConfig:
    properties:
      activeKeyID:
        type: string
      algorithm:
        $ref: '#/definitions/tenant.Algorithm'
    type: object
  tenant.Tenant:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      domain:
        type: string
      id:
        type: string
      issuer:
        type: string
      mfa:
        $ref: '#/definitions/tenant.MFAPolicy'
      settings:
        $ref: '#/definitions/tenant.TenantSettings'
      signingConfig:
        $ref: '#/definitions/tenant.SigningConfig'
      tokenVersion:
        type: integer
      updatedAt:
        type: string
    type: object
  tenant.TenantSettings:
    properties:
      access_token_ttl:
        description: seconds, default 3600 (1h)
        type: integer
      agent_rate_limit:
        description: requests per minute for M2M clients, 0 = use global
        type: integer
      allowed_origins:
        description: CORS per tenant
        items:
          type: string
        type: array
      lockout_duration:
        description: seconds
        type: integer
      max_login_attempts:
        description: 0 = unlimited
        type: integer
      password_min_length:
        description: default 8
        type: integer
      password_require_number:
        type: boolean
      password_require_special:
        type: boolean
      password_require_uppercase:
        type: boolean
      refresh_token_ttl:
        description: seconds, default 2592000 (30d)
        type: integer
      session_ttl:
        description: seconds, default 86400 (24h)
        type: integer
      smtp_from:
        type: string
      smtp_host:
        description: SMTP per-tenant override (empty = use global config)
        type: string
      smtp_password:
        type: string
      smtp_port:
        type: integer
      smtp_username:
        type: string
    type: object
  tenant.UpdateTenantRequest:
    properties:
      domain:
        type: string
      issuer:
        type: string
      mfa:
        $ref: '#/definitions/tenant.MFAPolicy'
      settings:
        $ref: '#/definitions/tenant.TenantSettings'
    type: object
...
info:
  contact: {}
paths:
...
  /tenants:
    get:
      parameters:
      - description: Pagination offset
        in: query
        name: offset
        type: integer
      - description: Page size (default 20)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/tenant.Tenant'
            type: array
      security:
      - AdminAuth: []
      summary: List tenants
      tags:
      - tenants
    post:
      consumes:
      - application/json
      parameters:
      - description: Tenant details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/tenant.CreateTenantRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/tenant.Tenant'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httputil.Error'
      security:
      - AdminAuth: []
      summary: Create tenant
      tags:
      - tenants
  /tenants/{id}:
    delete:
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.Error'
      security:
      - AdminAuth: []
      summary: Delete tenant
      tags:
      - tenants
    get:
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tenant.Tenant'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.Error'
      security:
      - AdminAuth: []
      summary: Get tenant
      tags:
      - tenants
    put:
      consumes:
      - application/json
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/tenant.UpdateTenantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tenant.Tenant'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.Error'
      security:
      - AdminAuth: []
      summary: Update tenant
      tags:
      - tenants
...
swagger: "2.0"